complete workflow example

+++ src/main/java/com/github/collabeng/api/error/PlanNotFoundException.java
@@ -6,7 +6,7 @@ import static java.lang.String.format;
  * Created by paul.smout on 09/03/2015.
  */
 public class PlanNotFoundException extends BusinessException{
-    public PlanNotFoundException(int id) {
+    public PlanNotFoundException(Long id) {
         super(format("Plan Not Found for id %d",id));
     }
 }
